pyproject.toml

[tool.poetry]
name = "aiocr"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9, !=3.9.7"
streamlit = "^1.23.1"
pandas = "^2.0.3"
pydantic = "^2.0.3"
opencv-python = "^4.8.0.74"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
pydantic-settings = "^2.10.1"
openai = "^1.98.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
